feat add some

+++ SyncRazorForms/Models/Product.cs
@@ -2,9 +2,9 @@
 
 public class Product
 {
-    public long Id { get; set; } 
+    public int Id { get; set; } 
     public string? Name { get; set; }
     public string? Description { get; set; }
     public double Cost { get; set; }
-    public long Amount { get; set; }
+    public int Amount { get; set; }
 }
